+++ Language/bili.JSONParser2.ja.po
@@ -1,0 +1,32 @@
+# bili - PrintInfo.py Strings
+# (C) 2019-2020 lifegpc
+# This file is distributed under the same license as the bili package.
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+msgid ""
+msgstr ""
+"Project-Id-Version: bili\n"
+"POT-Creation-Date: 2020-07-12 10:01:00+0000\n"
+"PO-Revision-Date: 2020-07-12 10:25+0000\n"
+"Language-Team: Japanese (https://www.transifex.com/lifegpc/teams/111556/ja/)\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Language: ja\n"
+"Plural-Forms: nplurals=1; plural=0;\n"
+
+msgctxt "OUTPUT1"
+msgid "Unable to get page <number> of Favorites. Try again now……"
+msgstr ""
+
+msgctxt "OUTPUT2"
+msgid "Unable to get page <number> of Channel. Try again now……"
+msgstr ""
+
+msgctxt "OUTPUT3"
+msgid "Unable to get page <number>. Try again now……"
+msgstr ""
+
+msgctxt "OUTPUT4"
+msgid "Unable to get uploader's information. Try again now……"
+msgstr ""
