+++ Language/bili.videodownload.ja.po
@@ -3,11 +3,15 @@
 # This file is distributed under the same license as the bili package.
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+# Translators:
+# Kum4423, 2020
+# 
 msgid ""
 msgstr ""
 "Project-Id-Version: bili\n"
 "POT-Creation-Date: 2020-07-11 12:54:00+0000\n"
 "PO-Revision-Date: 2020-07-11 13:31+0000\n"
+"Last-Translator: Kum4423, 2020\n"
 "Language-Team: Japanese (https://www.transifex.com/lifegpc/teams/111556/ja/)\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -21,7 +25,7 @@ msgstr ""
 
 msgctxt "OUTPUT2"
 msgid "Start downloading."
-msgstr ""
+msgstr "ダウンロードを開始します。"
 
 msgctxt "OUTPUT3"
 msgid ""
